fix fastify auth route cors handling
